be/requirements.txt
python-dotenv==1.0.1
requests==2.28.1
pymongo==4.11.2
flask==2.3.3
flask-cors==4.0.0
